requirements-backup.txt
astroid==2.8.2
autopep8==1.5.7
click==8.0.3
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.2
isort==5.9.3
itsdangerous==2.0.1
Jinja2==3.0.2
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
pkg_resources==0.0.0
platformdirs==2.4.0
psycopg2-binary==2.9.1
pycodestyle==2.8.0
pylint==2.11.1
SQLAlchemy==1.4.25
toml==0.10.2
typing-extensions==3.10.0.2
Werkzeug==2.0.2
wrapt==1.12.1
WTForms==2.3.3
python-cas==1.6.0
flask-apscheduler==1.5.0
flask-mail==0.9.1
Flask_Migrate-3.1.0
